LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/position-change-heuristic-in-nested-scroll-box-expected.txt


FAIL Position changes in document scroller. assert_equals: expected 225 but got 200
FAIL Position changes in scrollable <div>. assert_equals: expected 225 but got 200
